Adjust Paparazzi tests + custom tasks
